requirements.txt
openai
fastapi
uvicorn
openenv-core